requirements.txt
tensorflow==2.4.1
scipy==1.6.2
pandas==1.2.3
numpy==1.19.5
neptune-client==0.9.0
neptune-tensorflow-keras==0.9.0
librosa==0.8.0
jupyter==1.1.0
matplotlib==3.3.1
seaborn==0.11.0

